Fix doctests and remove orphan functions
fix
